+++ backend/controllers/provider.js
@@ -1,11 +1,12 @@
 const Provider = require('../models/provider');
 const clientService = require('../service/client');
 const { validationResult } = require('express-validator');
+const { BAD_REQUEST } = require('../const');
 
 exports.addProvider = async (req, res) => {
   const errors = validationResult(req);
   if (!errors.isEmpty()) {
-    return res.status(500).json(errors.array());
+    return res.status(BAD_REQUEST).json(errors.array());
   }
   const provider = new Provider(req.body.data);
   await provider.save();
@@ -20,7 +21,7 @@ exports.getProviders = async (req, res) => {
 exports.updateProvider = async (req, res) => {
   const errors = validationResult(req);
   if (!errors.isEmpty()) {
-    return res.status(500).json(errors.array());
+    return res.status(BAD_REQUEST).json(errors.array());
   }
   const provider = await Provider.findByIdAndUpdate(
     req.params.providerId,
@@ -31,6 +32,10 @@ exports.updateProvider = async (req, res) => {
 };
 
 exports.deleteProvider = async (req, res) => {
+  const errors = validationResult(req);
+  if (!errors.isEmpty()) {
+    return res.status(BAD_REQUEST).json(errors.array());
+  }
   await clientService.eraseProviderIdFromClients(req.params.providerId);
   const provider = await Provider.findByIdAndDelete(req.params.providerId);
   return res.json({ provider });
